+++ src/main/java/org/ojalgo/data/cluster/PointDistanceCache.java
@@ -1,14 +1,35 @@
+/*
+ * Copyright 1997-2025 Optimatika
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
 package org.ojalgo.data.cluster;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
-import java.util.function.ToDoubleBiFunction;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import org.ojalgo.array.ArrayR064;
 import org.ojalgo.array.NumberList;
+import org.ojalgo.data.proximity.DistanceMeasure;
 import org.ojalgo.function.constant.PrimitiveMath;
 import org.ojalgo.random.SampleSet;
 
@@ -32,6 +53,9 @@ final class PointDistanceCache {
         }
     }
 
+    /**
+     * Pick the centroid for this cluster (one of its current members)
+     */
     Point centroid(final Collection<Point> cluster) {
 
         Point retVal = null;
@@ -54,14 +78,24 @@ final class PointDistanceCache {
         return retVal;
     }
 
+    /**
+     * Get the distance between these two points.
+     */
     double distance(final Point point1, final Point point2) {
         return this.distance(point1.id, point2.id);
     }
 
+    /**
+     * The distance threshold used in {@link #initialiser(Collection)} to determine when the greedy clustering
+     * algorithm should not place a point in any of the existing clusters, but rather create a new cluster.
+     */
     double getThreshold() {
         return mySampleSet.swap(myValues).getMedian();
     }
 
+    /**
+     * Generate an initial set of centroids
+     */
     List<Point> initialiser(final Collection<Point> input) {
 
         GreedyClustering<Point> greedy = new GreedyClustering<>(this::centroid, this::distance, this.getThreshold());
@@ -77,7 +111,10 @@ final class PointDistanceCache {
         }).mapToObj(centroids::get).collect(Collectors.toList());
     }
 
-    void setup(final Collection<Point> input, final ToDoubleBiFunction<Point, Point> distanceCalculator) {
+    /**
+     * Set up the cache (configure this instance)
+     */
+    void setup(final Collection<Point> input, final DistanceMeasure measure) {
 
         int nbPoints = input.size();
 
@@ -95,7 +132,7 @@ final class PointDistanceCache {
             for (Point pointC : input) {
                 int col = pointC.id;
                 if (row > col) {
-                    myValues.add(myDistances[row][col] = distanceCalculator.applyAsDouble(pointR, pointC));
+                    myValues.add(myDistances[row][col] = pointR.distance(measure, pointC));
                 }
             }
         }